Move ProxyArtifactStore to mrm-servlet
