+++ mrm-api/pom.xml
@@ -31,8 +31,8 @@
 
     <!-- Maven core -->
     <dependency>
-      <groupId>org.eclipse.aether</groupId>
-      <artifactId>aether-api</artifactId>
+      <groupId>org.apache.maven.resolver</groupId>
+      <artifactId>maven-resolver-api</artifactId>
     </dependency>
     <dependency>
       <groupId>org.apache.maven</groupId>
