+++ pom.xml
@@ -110,9 +110,9 @@
         <scope>provided</scope>
       </dependency>
       <dependency>
-        <groupId>org.eclipse.aether</groupId>
-        <artifactId>aether-api</artifactId>
-        <version>1.1.0</version>
+        <groupId>org.apache.maven.resolver</groupId>
+        <artifactId>maven-resolver-api</artifactId>
+        <version>1.9.27</version>
         <scope>provided</scope>
       </dependency>
       <dependency>
